reine_des_maracas/requirements.txt
requests>=2.31.0
chromadb>=0.5.0 ; extra == "chroma"
fastapi>=0.111
uvicorn>=0.30
pydantic>=2
scikit-learn>=1.3.0
numpy>=1.24.0
vl-convert-python>=1.7.0

# Core Google Cloud & ADK dependencies
google-cloud-bigquery>=3.25.0
google-adk>=0.2.0
google-cloud-aiplatform>=1.114.0
google-genai>=0.5.0
